live data

+++ RecyclerviewViewModelLiveData/app/src/main/java/com/example/abc/recyclerviewviewmodellivedata/IDelete.java
@@ -1,5 +1,5 @@
 package com.example.abc.recyclerviewviewmodellivedata;
 
 public interface IDelete {
-    void delete(int position);
+    void delete(long position);
 }
